fix bound form fields replacing unbound form fields
This fixes a bug which causes a form submitted once to always be rendered
with any previous errors and data. b.fields was being created with a
shallow copy of f.fields, causing them to both reference the same object.
This commit fixes the issue by creating a new object for b and copying the
keys from f.